Login
I haven't implemented the isValidUsername (in login controller) part, because we dont have User objects yet. 

Need to serialize all User objects that exist in order to open them and see if valid username. But also idk how to refer to specific user n use UserController based on that user. For starters, our list of Users will begin by only having 'stock' user with it's photos we make that manually not through program.